upload multiple memes @ once
