+++ packages/web/components/CreateMeme/CreateMeme.tsx
@@ -1,7 +1,9 @@
+import { EditorContent } from "@tiptap/react";
 import { observer } from "mobx-react-lite";
 import Image from "next/image";
 import { ReactNode, useMemo } from "react";
 import { useDropzone } from "react-dropzone";
+import { AiOutlineMinus, AiOutlinePlus } from "react-icons/ai";
 import { IoCloseOutline } from "react-icons/io5";
 import {
   RedditIcon,
@@ -13,13 +15,19 @@ import { getBaseUrl } from "../../environment/vars";
 import { objectKeys } from "../../helpers/arrays";
 import { css } from "../../helpers/css";
 import AppStore from "../../store/App.store";
-import CreateMemeStore, { CreateMemeView } from "../../store/CreateMeme.store";
+import CreateMemeStore, {
+  CreateMemeView,
+  MemeStore,
+} from "../../store/CreateMeme.store";
 import AspectRatio from "../DSL/AspectRatio/AspectRatio";
+import Button from "../DSL/Button/Button";
 import Form from "../DSL/Form/Form";
-import { FileWithPreview } from "../DSL/Form/MediaInput";
+import { FormLabel } from "../DSL/Form/FormControl";
 import TextInput from "../DSL/Form/TextInput";
 import Pane from "../DSL/Pane/Pane";
 import Text, { TextSize, TextType } from "../DSL/Text/Text";
+import { useTipTapEditor } from "../TipTapEditor/TipTapEditor";
+import TipTapEditorToolbar from "../TipTapEditor/TipTapEditorToolbar";
 
 const CreateMeme: React.FC<{
   store: CreateMemeStore;
@@ -85,15 +93,14 @@ const CreateMemeForm: React.FC<{
         </div>
       </Form> */}
       <div className={css("flex", "flex-col", "gap-4")}>
-        {store.hasFiles && (
+        {store.hasMemes && (
           <div
             className={css("grid", "grid-cols-1", "md:grid-cols-2", "gap-4")}
           >
-            {store.files.map((file, index) => (
+            {store.memes.map((memeStore, index) => (
               <MemeForm
                 key={`meme-form-input-${index}`}
-                store={store}
-                file={file}
+                store={memeStore}
                 onRemove={() => store.removeFile(index)}
               />
             ))}
@@ -104,7 +111,7 @@ const CreateMemeForm: React.FC<{
             store={store}
             maxSizeBites={AppStore.settings.maxSizeBytes}
             acceptedMimeToExtension={AppStore.settings.mimeTypeToExtension}
-            title={store.hasFiles ? "+ Add more" : "Drop memes for money"}
+            title={store.hasMemes ? "+ Add more" : "Drop memes for money"}
           />
         )}
       </div>
@@ -113,16 +120,19 @@ const CreateMemeForm: React.FC<{
 });
 
 interface MemeFormProps {
-  file: FileWithPreview;
-  store: CreateMemeStore;
+  store: MemeStore;
   onRemove: () => void;
 }
 
-const MemeForm = ({ file, onRemove }: MemeFormProps) => {
+const MemeForm = observer(({ store, onRemove }: MemeFormProps) => {
+  const editor = useTipTapEditor(
+    store.description ? store.description : "",
+    true
+  );
   return (
     <Pane
-      key={file.name}
-      // title={}
+      key={store.file.name}
+      title={store.name}
       rightOfTitle={
         <button
           onClick={() => onRemove()}
@@ -136,18 +146,73 @@ const MemeForm = ({ file, onRemove }: MemeFormProps) => {
     >
       <div className={css("relative", "w-full", "h-[300px]")}>
         <Image
-          alt={file.name}
-          src={file.preview}
+          alt={store.file.name}
+          src={store.file.preview}
           fill
           className={css("object-contain")}
         />
       </div>
-      <Form onSubmit={async () => {}}>
-        <TextInput name={`text-input`} label={"Name"} block />
+      <Form
+        onSubmit={async () => {}}
+        className={css("flex", "flex-col", "gap-2", "mt-2", "items-start")}
+      >
+        {store.showName && (
+          <div className={css("flex", "items-center", "gap-2", "w-full")}>
+            <TextInput
+              leftOfInput={
+                <Button onClick={() => store.toggleShowName()}>
+                  <AiOutlineMinus size={12} />
+                </Button>
+              }
+              block
+              name={`text-input`}
+              label={"Name"}
+              value={store.name}
+              onChange={(value) => (store.name = value)}
+            />
+          </div>
+        )}
+        {!store.showName && (
+          <Button onClick={() => store.toggleShowName()}>
+            <div className={css("flex", "items-center", "gap-0.5")}>
+              <AiOutlinePlus size={15} />
+              <Text>name</Text>
+            </div>
+          </Button>
+        )}
+        {store.showDescription && (
+          <div className={css("flex", "items-center", "gap-2", "w-full")}>
+            <div className={css("grow")}>
+              <FormLabel>Description</FormLabel>
+              <div className={css("flex", "items-start", "gap-2")}>
+                <div className={css("mt-1")}>
+                  <Button onClick={() => store.toggleShowDescription()}>
+                    <AiOutlineMinus size={12} />
+                  </Button>
+                </div>
+                <div className={css("grow")}>
+                  <EditorContent editor={editor} />
+                </div>
+              </div>
+              <TipTapEditorToolbar
+                store={store.toolbarStore}
+                editor={editor!}
+              />
+            </div>
+          </div>
+        )}
+        {!store.showDescription && (
+          <Button onClick={() => store.toggleShowDescription()}>
+            <div className={css("flex", "items-center", "gap-0.5")}>
+              <AiOutlinePlus size={15} />
+              <Text>description</Text>
+            </div>
+          </Button>
+        )}
       </Form>
     </Pane>
   );
-};
+});
 
 interface MemeInputProps {
   store: CreateMemeStore;
@@ -161,6 +226,8 @@ const MemeInput = observer(
     const { getRootProps, getInputProps, isDragActive } = useDropzone({
       onDropAccepted: (files) => store.onDropAccepted(files),
       multiple: true,
+      maxSize: maxSizeBites,
+      maxFiles: 100,
     });
     const acceptedExtensionsLabel = useMemo(() => {
       const extensions = objectKeys(acceptedMimeToExtension).reduce(
